+++ src/Core/Model/Subscription/AzureServiceBusDestination.php
@@ -14,10 +14,6 @@ use Commercetools\Core\Model\Common\Context;
  * @method AzureServiceBusDestination setType(string $type = null)
  * @method string getConnectionString()
  * @method AzureServiceBusDestination setConnectionString(string $connectionString = null)
- * @method string getUri()
- * @method AzureServiceBusDestination setUri(string $uri = null)
- * @method string getAccessKey()
- * @method AzureServiceBusDestination setAccessKey(string $accessKey = null)
  */
 class AzureServiceBusDestination extends Destination
 {
@@ -26,13 +22,12 @@ class AzureServiceBusDestination extends Destination
         return [
             'type' => [static::TYPE => 'string'],
             'connectionString' => [static::TYPE => 'string'],
-            'uri' => [static::TYPE => 'string'],
-            'accessKey' => [static::TYPE => 'string'],
         ];
     }
 
     /**
-     * @param $connectionString
+     * @deprecated use ofConnectionString instead
+     * @param string $connectionString
      * @param Context|null $context
      * @return AzureServiceBusDestination
      */
@@ -44,16 +39,14 @@ class AzureServiceBusDestination extends Destination
     }
 
     /**
-     * @param $uri
-     * @param $accessKey
+     * @param string $connectionString
      * @param Context|null $context
      * @return AzureServiceBusDestination
      */
-    public static function ofUriAndAccessKey($uri, $accessKey, Context $context = null)
+    public static function ofConnectionString($connectionString, Context $context = null)
     {
         return static::of($context)
             ->setType(static::DESTINATION_AZURE_SERVICE_BUS)
-            ->setUri($uri)
-            ->setAccessKey($accessKey);
+            ->setConnectionString($connectionString);
     }
 }
