+++ src/Core/Model/Subscription/Destination.php
@@ -18,6 +18,7 @@ class Destination extends JsonObject
     const DESTINATION_SQS = 'SQS';
     const DESTINATION_IRON_MQ = 'IronMQ';
     const DESTINATION_SNS = 'SNS';
+    const DESTINATION_AZURE_EVENT_GRID = 'EventGrid';
     const DESTINATION_AZURE_SERVICE_BUS = 'AzureServiceBus';
     const DESTINATION_GOOGLE_CLOUD_PUB_SUB = 'GoogleCloudPubSub';
 
@@ -34,6 +35,7 @@ class Destination extends JsonObject
             static::DESTINATION_SQS => SQSDestination::class,
             static::DESTINATION_IRON_MQ => IronMQDestination::class,
             static::DESTINATION_SNS => SNSDestination::class,
+            static::DESTINATION_AZURE_EVENT_GRID => AzureEventGridDestination::class,
             static::DESTINATION_AZURE_SERVICE_BUS => AzureServiceBusDestination::class,
             static::DESTINATION_GOOGLE_CLOUD_PUB_SUB => GoogleCloudPubSubDestination::class,
         ];
